Split Admin Custom Sounds components
