+++ client/admin/customSounds/AdminSounds.js
@@ -1,24 +1,12 @@
-import React, { useMemo, useCallback, useState, useEffect } from 'react';
-import { Box, Table, TextInput, Icon, Button } from '@rocket.chat/fuselage';
+import React, { useMemo, useCallback } from 'react';
+import { Box, Table, Icon, Button } from '@rocket.chat/fuselage';
 
 import { useTranslation } from '../../contexts/TranslationContext';
 import { GenericTable, Th } from '../../components/GenericTable';
 import { useCustomSound } from '../../contexts/CustomSoundContext';
+import FilterByText from './FilterByText';
 
-const FilterByText = ({ setFilter, ...props }) => {
-	const t = useTranslation();
-	const [text, setText] = useState('');
-	const handleChange = useCallback((event) => setText(event.currentTarget.value), []);
-
-	useEffect(() => {
-		setFilter({ text });
-	}, [text]);
-	return <Box mb='x16' is='form' onSubmit={useCallback((e) => e.preventDefault(), [])} display='flex' flexDirection='column' {...props}>
-		<TextInput flexShrink={0} placeholder={t('Search')} addon={<Icon name='magnifier' size='x20'/>} onChange={handleChange} value={text} />
-	</Box>;
-};
-
-export function AdminSounds({
+function AdminSounds({
 	data,
 	sort,
 	onClick,
@@ -54,3 +42,5 @@ export function AdminSounds({
 
 	return <GenericTable FilterComponent={FilterByText} header={header} renderRow={renderRow} results={data.sounds} total={data.total} setParams={setParams} params={params} />;
 }
+
+export default AdminSounds;
