+++ client/admin/customSounds/AdminSoundsRoute.js
@@ -7,9 +7,9 @@ import { useDebouncedValue } from '@rocket.chat/fuselage-hooks';
 import { usePermission } from '../../contexts/AuthorizationContext';
 import { useTranslation } from '../../contexts/TranslationContext';
 import Page from '../../components/basic/Page';
-import { AdminSounds } from './AdminSounds';
-import { AddCustomSound } from './AddCustomSound';
-import { EditCustomSound } from './EditCustomSound';
+import AdminSounds from './AdminSounds';
+import AddCustomSound from './AddCustomSound';
+import EditCustomSound from './EditCustomSound';
 import { useRoute, useRouteParameter } from '../../contexts/RouterContext';
 import { useEndpointData } from '../../hooks/useEndpointData';
 import VerticalBar from '../../components/basic/VerticalBar';
